Add joinpath to BucketFSLocations and moved some tests from integration tests to unit tests
